+++ src/py_rizmi/__init__.py
@@ -9,7 +9,16 @@ from py_rizmi.core.hwid import HardwareIdentifier as MachineFingerprint
 from py_rizmi.core.keypair import KeyPairManager as KeyPair
 from py_rizmi.core.license_issuer import LicenseIssuer
 from py_rizmi.core.license_validator import LicenseValidator
+from py_rizmi.core.swap_auth import (
+    create_replacement_authorization_payload,
+    create_swap_request,
+    sign_authorization_payload,
+    sign_swap_request,
+    verify_authorization,
+    verify_swap_authorization,
+)
 from py_rizmi.models.license_payload import LicensePayload
+from py_rizmi.models.swap_payload import LicenseSwapPayload, ReplacementAuthorizationPayload
 
 __all__ = [
     "__version__",
@@ -18,4 +27,12 @@ __all__ = [
     "KeyPair",
     "MachineFingerprint",
     "LicensePayload",
+    "LicenseSwapPayload",
+    "ReplacementAuthorizationPayload",
+    "create_swap_request",
+    "create_replacement_authorization_payload",
+    "sign_swap_request",
+    "sign_authorization_payload",
+    "verify_swap_authorization",
+    "verify_authorization",
 ]