+++ src/py_rizmi/models/__init__.py
@@ -1,7 +1,10 @@
-"""Data models for license payloads."""
+"""Data models for license payloads and license swap payloads."""
 
 from py_rizmi.models.license_payload import LicensePayload
+from py_rizmi.models.swap_payload import LicenseSwapPayload, ReplacementAuthorizationPayload
 
 __all__ = [
     "LicensePayload",
+    "LicenseSwapPayload",
+    "ReplacementAuthorizationPayload",
 ]